@@ -29,50 +29,52 @@ class Module extends \Icybee\Module
 	 */
 	protected function lazy_get_views()
 	{
-		return \ICanBoogie\array_merge_recursive
-		(
-			parent::lazy_get_views(), array
-			(
-				'view' => array
-				(
-					'title' => "Record detail",
-					'provider' => __NAMESPACE__ . '\ViewProvider',
-					'assets' => array(),
-					'renders' => \Icybee\Modules\Views\View::RENDERS_ONE
-				),
+		return \ICanBoogie\array_merge_recursive(parent::lazy_get_views(), [
 
-				'list' => array
-				(
-					'title' => "Records list",
-					'provider' => __NAMESPACE__ . '\ViewProvider',
-					'assets' => array(),
-					'renders' => \Icybee\Modules\Views\View::RENDERS_MANY
-				),
+			'view' => [
 
-				'home' => array
-				(
-					'title' => "Records home",
-					'provider' => __NAMESPACE__ . '\ViewProvider',
-					'assets' => array(),
-					'renders' => \Icybee\Modules\Views\View::RENDERS_MANY
-				)
-			)
-		);
+				'title' => "Record detail",
+				'provider' => __NAMESPACE__ . '\ViewProvider',
+				'assets' => [],
+				'renders' => \Icybee\Modules\Views\View::RENDERS_ONE
+
+			],
+
+			'list' => [
+
+				'title' => "Records list",
+				'provider' => __NAMESPACE__ . '\ViewProvider',
+				'assets' => [],
+				'renders' => \Icybee\Modules\Views\View::RENDERS_MANY
+
+			],
+
+			'home' => [
+
+				'title' => "Records home",
+				'provider' => __NAMESPACE__ . '\ViewProvider',
+				'assets' => [],
+				'renders' => \Icybee\Modules\Views\View::RENDERS_MANY
+
+			]
+
+		]);
 	}
 
 	protected function resolve_primary_model_tags($tags)
 	{
-		return parent::resolve_model_tags($tags, 'primary') + array
-		(
+		return parent::resolve_model_tags($tags, 'primary') + [
+
 			Model::T_CONSTRUCTOR => $this->id
-		);
+
+		];
 	}
 
 	static public function create_default_routes()
 	{
 		global $core;
 
-		$routes = array();
+		$routes = [];
 		$modules = $core->modules;
 
 		foreach ($modules->enabled_modules_descriptors as $module_id => $descriptor)
@@ -82,81 +84,77 @@ class Module extends \Icybee\Module
 				continue;
 			}
 
-			$common = array
-			(
+			$common = [
+
 				'module' => $module_id,
 				'controller' => 'Icybee\BlockController',
 				'visibility' => 'visible'
-			);
+
+			];
 
 // 			\ICanBoogie\log("create default routes for $module_id");
 
 			# manage (index)
 
-			$routes["admin:$module_id"] = array
-			(
+			$routes["admin:$module_id"] = [
+
 				'pattern' => "/admin/$module_id",
 				'title' => '.manage',
 				'block' => 'manage',
 				'index' => true
-			)
 
-			+ $common;
+			] + $common;
 
 			if ($module_id == 'contents' || $modules->is_extending($module_id, 'contents') || $module_id == 'files' || $modules->is_extending($module_id, 'files'))
 			{
 				# config'
 
-				$routes["admin:$module_id/config"] = array
-				(
+				$routes["admin:$module_id/config"] = [
+
 					'pattern' => "/admin/$module_id/config",
 					'title' => '.config',
 					'block' => 'config',
 					'permission' => self::PERMISSION_ADMINISTER,
-				)
 
-				+ $common;
+				] + $common;
 			}
 
 			# create
 
-			$routes["admin:$module_id/new"] = array
-			(
+			$routes["admin:$module_id/new"] = [
+
 				'pattern' => "/admin/$module_id/new",
 				'title' => '.new',
 				'block' => 'edit'
-			)
 
-			+ $common;
+			] + $common;
 
 			# edit
 
-			$routes["admin:$module_id/edit"] = array
-			(
+			$routes["admin:$module_id/edit"] = [
+
 				'pattern' => "/admin/$module_id/<\d+>/edit",
 				'controller' => 'Icybee\EditController',
 				'title' => '.edit',
 				'block' => 'edit',
 				'visibility' => 'auto'
-			)
 
-			+ $common;
+			] + $common;
 
 			# delete
 
-			$routes["admin:$module_id/delete"] = array
-			(
+			$routes["admin:$module_id/delete"] = [
+
 				'pattern' => "/admin/$module_id/<\d+>/delete",
 				'controller' => 'Icybee\DeleteController',
 				'title' => '.delete',
 				'block' => 'delete',
 				'visibility' => 'auto'
-			)
 
-			+ $common;
+			] + $common;
 		}
 
-		new Module\CreateDefaultRoutesEvent($modules['nodes'], array('routes' => &$routes));
+		new Module\CreateDefaultRoutesEvent($modules['nodes'], [ 'routes' => &$routes ]);
 
 // 		var_dump($routes);
 
