@@ -4,19 +4,21 @@ namespace Icybee\Modules\Nodes;
 
 $hooks = __NAMESPACE__ . '\Hooks::';
 
-return array
-(
-	'system-nodes-now' => array
-	(
+return [
+
+	'system-nodes-now' => [
+
 		'title' => "From a glance",
 		'callback' => $hooks . 'dashboard_now',
 		'column' => 0
-	),
 
-	'system-nodes-user-modified' => array
-	(
+	],
+
+	'system-nodes-user-modified' => [
+
 		'title' => "Your last modifications",
 		'callback' => $hooks . 'dashboard_user_modified',
 		'column' => 0
-	)
-);
+	]
+
+];
