@@ -5,48 +5,55 @@ namespace Icybee\Modules\Nodes;
 use ICanBoogie\ActiveRecord\Model;
 use ICanBoogie\Module;
 
-return array
-(
+return [
+
 	Module::T_CATEGORY => 'contents',
 	Module::T_DESCRIPTION => 'Centralized node system base',
-	Module::T_MODELS => array
-	(
-		'primary' => array
-		(
-			Model::SCHEMA => array
-			(
-				'fields' => array
-				(
+	Module::T_MODELS => [
+
+		'primary' => [
+
+			Model::SCHEMA => [
+
+				'fields' => [
+
 					'nid' => 'serial',
 					'uid' => 'foreign',
 					'siteid' => 'foreign',
 					'nativeid' => 'foreign',
-					'constructor' => array('varchar', 64, 'indexed' => true),
+					'constructor' => [ 'varchar', 64, 'indexed' => true ],
 					'title' => 'varchar',
-					'slug' => array('varchar', 80, 'indexed' => true),
-					'language' => array('varchar', 8, 'indexed' => true),
-					'created_at' => array('timestamp', 'default' => 'CURRENT_TIMESTAMP'),
+					'slug' => [ 'varchar', 80, 'indexed' => true ],
+					'language' => [ 'varchar', 8, 'indexed' => true ],
+					'created_at' => [ 'timestamp', 'default' => 'CURRENT_TIMESTAMP' ],
 					'updated_at' => 'timestamp',
-					'is_online' => array('boolean', 'indexed' => true)
-				)
-			)
-		)
-	),
+					'is_online' => [ 'boolean', 'indexed' => true ]
+
+				]
+
+			]
+
+		]
+
+	],
 
 	Module::T_NAMESPACE => __NAMESPACE__,
 	Module::T_PERMISSION => false,
-	Module::T_PERMISSIONS => array
-	(
+	Module::T_PERMISSIONS => [
+
 		'modify belonging site'
-	),
+
+	],
 
 	Module::T_REQUIRED => true,
-	Module::T_REQUIRES => array
-	(
+	Module::T_REQUIRES => [
+
 		'sites' => '2.0',
 		'users' => '2.0'
-	),
+
+	],
 
 	Module::T_TITLE => 'Nodes',
 	Module::T_VERSION => '2.0'
-);
+
+];
