@@ -18,19 +18,16 @@ class OnlineOperation extends Operation
 {
 	/**
 	 * Controls for the operation: permission(maintain), record and ownership.
-	 *
-	 * @see ICanBoogie.Operation::get_controls()
 	 */
 	protected function get_controls()
 	{
-		return array
-		(
+		return [
+
 			self::CONTROL_PERMISSION => Module::PERMISSION_MAINTAIN,
 			self::CONTROL_RECORD => true,
 			self::CONTROL_OWNERSHIP => true
-		)
 
-		+ parent::get_controls();
+		] + parent::get_controls();
 	}
 
 	protected function validate(\ICanboogie\Errors $errors)
@@ -40,8 +37,6 @@ class OnlineOperation extends Operation
 
 	/**
 	 * Changes the target record is_online property to true and updates the record.
-	 *
-	 * @see ICanBoogie.Operation::process()
 	 */
 	protected function process()
 	{
@@ -49,7 +44,7 @@ class OnlineOperation extends Operation
 		$record->is_online = true;
 		$record->save();
 
-		$this->response->message = new FormattedString('!title is now online', array('!title' => $record->title));
+		$this->response->message = new FormattedString('!title is now online', [ '!title' => $record->title ]);
 
 		return true;
 	}
